MOVE - 테스트 디렉토리에 racing 패키지 추가 후 관련 테스트들을 옮김
