+++ src/test/java/study/racing/CarRacingTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.racing;
 
 public class CarRacingTest {
 }
