+++ src/test/java/study/racing/CarTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.racing;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
